Split up functional AppTest in multiple tests
