+++ tests/Functional/AppTest.php
@@ -2,14 +2,29 @@
 
 namespace Starch\Tests\Functional;
 
-use Interop\Http\ServerMiddleware\DelegateInterface;
-use Psr\Http\Message\ServerRequestInterface;
 use Starch\Router\RouterMiddleware;
 use Starch\Tests\AppTestCase;
 use Zend\Diactoros\Response;
+use Zend\Diactoros\Response\TextResponse;
 
 class AppTest extends AppTestCase
 {
+    public function testProcessesRequest()
+    {
+        $this->app->get('/', function() {
+            $response = new TextResponse('foo');
+
+            return $response;
+        });
+
+        $this->app->add(RouterMiddleware::class);
+
+        $response = $this->get('/');
+
+        $this->assertEquals(200, $response->getStatusCode());
+        $this->assertEquals('foo', (string)$response->getBody());
+    }
+
     public function testReturns404ResponseOnNotFound()
     {
         $response = $this->get('/');
@@ -17,7 +32,7 @@ class AppTest extends AppTestCase
         $this->assertEquals(404, $response->getStatusCode());
     }
 
-    public function testReturns404ResponseOnMethodNotAllowed()
+    public function testReturns405ResponseOnMethodNotAllowed()
     {
         $this->app->get('/', function() {
             return new Response();
@@ -26,91 +41,5 @@ class AppTest extends AppTestCase
         $response = $this->post('/');
 
         $this->assertEquals(405, $response->getStatusCode());
-    }
-
-    public function testAcceptsMiddleware()
-    {
-        $this->app->get('/', function(ServerRequestInterface $request) {
-            $response = new Response();
-            $response->getBody()->write($request->getHeader('x-name')[0]);
-
-            return $response;
-        });
-
-        $this->app->add(function (ServerRequestInterface $request, DelegateInterface $delegate) {
-            $request = $request->withHeader('x-name', 'foo');
-
-            return $delegate->process($request);
-        });
-        $this->app->add(function (ServerRequestInterface $request, DelegateInterface $delegate) {
-            $response = $delegate->process($request);
-            $response->getBody()->write('bar');
-
-            return $response;
-        });
-        $this->app->add(RouterMiddleware::class);
-
-        $response = $this->get('/');
-
-        $this->assertEquals(200, $response->getStatusCode());
-        $this->assertEquals('foobar', (string)$response->getBody());
-    }
-
-    public function testConstrainsMiddlewareToPath()
-    {
-        $this->app->get('/foo', function() {
-            return new Response();
-        });
-
-        $this->app->add(function (ServerRequestInterface $request, DelegateInterface $delegate) {
-            $response =  $delegate->process($request);
-
-            return $response->withHeader('x-foo', 'foo');
-        }, '/foo');
-
-        $this->app->add(function (ServerRequestInterface $request, DelegateInterface $delegate) {
-            $response =  $delegate->process($request);
-
-            return $response->withHeader('x-bar', 'foo');
-        }, '/bar');
-        $this->app->add(RouterMiddleware::class);
-
-        $response = $this->get('/foo');
-
-        $this->assertEquals(200, $response->getStatusCode());
-        $this->assertTrue($response->hasHeader('x-foo'));
-        $this->assertFalse($response->hasHeader('x-bar'));
-    }
-
-    public function testSetsRouteArguments()
-    {
-        $this->app->get('/{name}', function(ServerRequestInterface $request, $name) {
-            $response = new Response();
-            $response->getBody()->write($name);
-
-            return $response;
-        });
-        $this->app->add(RouterMiddleware::class);
-
-        $response = $this->get('/foo');
-
-        $this->assertEquals(200, $response->getStatusCode());
-        $this->assertEquals('foo', (string)$response->getBody());
-    }
-
-    public function testTestCaseCanSendPost()
-    {
-        $this->app->post('/', function(ServerRequestInterface $request) {
-            $response = new Response();
-            $response->getBody()->write($request->getParsedBody()['name']);
-
-            return $response;
-        });
-        $this->app->add(RouterMiddleware::class);
-
-        $response = $this->post('/', ['name' => 'foo']);
-
-        $this->assertEquals(200, $response->getStatusCode());
-        $this->assertEquals('foo', (string)$response->getBody());
     }
 }
